// src/models/profile-request/update-profile-info-request.ts
export type UpdateProfileInfoRequest = {
    phone_number: string;
    first_name: string;
    middle_name: string;
    last_name: string;
    date_of_birth: Date;
    hashtag_name: string;
    gender: string;
    language: string;
    country: string;
    profile_description: string;
    avatar_image: string;
    cover_image: string;
    job_name: string;
    time_zone: string;
}